pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.5.2",
    "dj-database-url>=2.3.0",
    "django>=5.2.1",
    "django-celery-beat>=2.8.0",
    "django-ckeditor>=6.7.2",
    "django-import-export>=4.3.7",
    "google-api-python-client>=2.169.0",
    "google-auth>=2.40.1",
    "google-auth-oauthlib>=1.2.2",
    "psycopg2-binary>=2.9.10",
    "python-decouple>=3.8",
    "redis>=6.0.0",
    "sendgrid>=6.12.0",
]
